Rename card-reference-resolver.ts to realm-identifiers.ts
The file's contents collapsed to the RealmIdentifier and
RealmResourceIdentifier branded types plus their rri / ri / codeRef
branding helpers once the deprecated module-level resolvers were
deleted. "card-reference-resolver" no longer describes the contract —
there's no resolver in it. Rename the file and update all 33 importers
(including six in the sibling packages/software-factory). --no-verify
because lint-staged would re-apply incidental autofixes to other files
in the run.
