+++ packages/realm-server/tests/index.ts
@@ -268,7 +268,7 @@ const ALL_TEST_FILES: string[] = [
   './boxel-domain-availability-test',
   './get-boxel-claimed-domain-test',
   './claim-boxel-domain-test',
-  './card-reference-resolver-test',
+  './realm-identifiers-test',
   './bfm-card-references-test',
   './package-shim-handler-test',
   './command-parsing-utils-test',
