+++ packages/runtime-common/tests/search-in-flight-key-test.ts
@@ -1,4 +1,4 @@
-import type { RealmResourceIdentifier } from '../card-reference-resolver';
+import type { RealmResourceIdentifier } from '../realm-identifiers';
 import type { SharedTests } from '../helpers';
 import type { Query } from '../query';
 import { searchInFlightKey } from '../realm-index-query-engine';
