+++ packages/software-factory/src/darkfactory-schemas.ts
@@ -13,7 +13,7 @@ import type {
   LooseSingleCardDocument,
   Relationship,
 } from '@cardstack/runtime-common';
-import { rri } from '@cardstack/runtime-common/card-reference-resolver';
+import { rri } from '@cardstack/runtime-common/realm-identifiers';
 
 import { logger } from './logger';
